Software Release CAIP
A proposal on how we can lay a solid foundation for requesting chain software information alongside a standard data structure for response payloads.
The hope is that this can lay a foundation on which systems can be built that address ongoing issues for node runners and validator operators know about and automatically executing on software release events.