improvements from @lukaisailovic feedback
This is useful so that we don't imply that the user always have to be prompted and select this each time. For example, if the wallet already has USDC in one account and that's the requested token then the address/token used can be pre-populated to make the UX prompt require just clicking through to approve it.